requirements.txt
streamlit==1.22.0
transformers==4.30.2
torch==2.0.1
PyPDF2==3.0.1
openai==0.27.8
python-dotenv==1.0.0